+++ src/material-table.js
@@ -410,6 +410,13 @@ export default class MaterialTable extends React.Component {
     );
   };
 
+  onGroupSelected = (checked, path) => {
+    this.dataManager.changeGroupSelected(checked, path);
+    this.setState(this.dataManager.getRenderState(), () =>
+      this.onSelectionChange()
+    );
+  };
+
   onChangeColumnHidden = (column, hidden) => {
     this.dataManager.changeColumnHidden(column, hidden);
     this.setState(this.dataManager.getRenderState(), () => {
@@ -943,6 +950,7 @@ export default class MaterialTable extends React.Component {
             ).length > 0
           }
           showSelectAllCheckbox={props.options.showSelectAllCheckbox}
+          showSelectGroupCheckbox={props.options.showSelectGroupCheckbox}
           orderBy={this.state.orderBy}
           orderDirection={this.state.orderDirection}
           onAllSelected={this.onAllSelected}
@@ -980,6 +988,7 @@ export default class MaterialTable extends React.Component {
         isTreeData={this.props.parentChildData !== undefined}
         onFilterChanged={this.onFilterChange}
         onRowSelected={this.onRowSelected}
+        onGroupSelected={this.onGroupSelected}
         onToggleDetailPanel={this.onToggleDetailPanel}
         onGroupExpandChanged={this.onGroupExpandChanged}
         onTreeExpandChanged={this.onTreeExpandChanged}
@@ -1037,9 +1046,10 @@ export default class MaterialTable extends React.Component {
       result.push(selectionWidth + 'px');
     }
 
-    for (let i = 0; i < Math.abs(count) && i < props.columns.length; i++) {
-      const colDef =
-        props.columns[count >= 0 ? i : props.columns.length - 1 - i];
+    for (let i = 0; i < Math.abs(count) && i < this.state.columns.length; i++) {
+      const colDef = this.state.columns[
+        count >= 0 ? i : this.state.columns.length - 1 - i
+      ];
       if (colDef.tableData) {
         if (typeof colDef.tableData.width === 'number') {
           result.push(colDef.tableData.width + 'px');