+++ types/index.d.ts
@@ -367,6 +367,7 @@ export interface Options<RowData extends object> {
   showEmptyDataSourceMessage?: boolean;
   showFirstLastPageButtons?: boolean;
   showSelectAllCheckbox?: boolean;
+  showSelectGroupCheckbox?: boolean;
   showTitle?: boolean;
   showTextRowsSelected?: boolean;
   showDetailPanelIcon?: boolean;