+++ CHANGELOG.md
@@ -262,6 +262,18 @@ We updated to the stable release of MUI v5. Please update accordingly.
 
 - # //github.com/villuv/material-table into villuv-feature/mui5 ([1a6d73ed](https://github.com/material-table-core/core/commit/1a6d73ed215a294f8dd66addf678a3267a70815a))
 
+#### 4.3.3 (2021-09-22)
+
+#### 4.3.2 (2021-09-22)
+
+##### New Features
+
+- Add group selection. ([#343](https://github.com/material-table-core/core/pull/343)) ([479798c2](https://github.com/material-table-core/core/commit/479798c28f63fe9c46b5f07b711362aff8c8d30a))
+
+##### Bug Fixes
+
+- regression of optional detail panel function ([6d55e366](https://github.com/material-table-core/core/commit/6d55e3660aa869df555197f1cf8bbfdf33b15d7f))
+
 #### 4.3.1 (2021-08-31)
 
 ##### Chores